Uprawnienia

+++ RestaurantAPI/Services/RestaurantService.cs
@@ -1,5 +1,9 @@
-﻿using AutoMapper;
+﻿using System.Security.Claims;
+using AutoMapper;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Models;
@@ -10,41 +14,55 @@ public interface IRestaurantService
 {
     RestaurantDto GetById(int id);
     IEnumerable<RestaurantDto> GetAll();
-    int Create(CreateRestaurantDto dto);
-    void Delete(int id);
-    void Update(int id, UpdateRestaurantDto dto);
+    int Create(CreateRestaurantDto dto, int userId);
+    void Delete(int id, ClaimsPrincipal user);
+    void Update(int id, UpdateRestaurantDto dto, ClaimsPrincipal user);
 }
 public class RestaurantService: IRestaurantService
 {
     private readonly RestaurantDbContext _dbContext;
     private readonly IMapper _mapper;
     private readonly ILogger<RestaurantService> _logger;
+    private readonly IAuthorizationService _authorizationService;
 
-    public RestaurantService(RestaurantDbContext dbContext, IMapper mapper,ILogger<RestaurantService> logger)
+    public RestaurantService(RestaurantDbContext dbContext, IMapper mapper,ILogger<RestaurantService> logger, IAuthorizationService authenticationService)
     {
         _dbContext = dbContext;
         _mapper = mapper;
         _logger = logger;
+        _authorizationService = authenticationService;
     }
 
-    public void Update(int id, UpdateRestaurantDto dto)
+    public void Update(int id, UpdateRestaurantDto dto, ClaimsPrincipal user)
     {
         var restaurant = _dbContext
             .Restaurants
             .FirstOrDefault(r => r.Id == id);
         if (restaurant is null) throw new NotFoundException("Restaurant not found");
+        var authorizationResult = _authorizationService.AuthorizeAsync(user, restaurant,
+            new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+        if (!authorizationResult.Succeeded)
+        {
+            throw new ForbidException();
+        }
         restaurant.Name = dto.Name;
         restaurant.Description = dto.Description;
         restaurant.HasDelivery = dto.HasDelivery;
         _dbContext.SaveChanges();
     }
-    public void Delete(int id)
+    public void Delete(int id, ClaimsPrincipal user)
     {
         _logger.LogError($"Restaurant with id: {id} DELETE action invoked");
         var restaurant = _dbContext
             .Restaurants
             .FirstOrDefault(r => r.Id == id);
         if (restaurant is null) throw new NotFoundException("Restaurant not found");
+        var authorizationResult = _authorizationService.AuthorizeAsync(user, restaurant,
+            new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+        if (!authorizationResult.Succeeded)
+        {
+            throw new ForbidException();
+        }
         _dbContext.Restaurants.Remove(restaurant);
         _dbContext.SaveChanges();
     }
@@ -71,9 +89,10 @@ public class RestaurantService: IRestaurantService
         return restaurantsDtos;
     }
 
-    public int Create(CreateRestaurantDto dto)
+    public int Create(CreateRestaurantDto dto, int userId)
     {
         var restaurant = _mapper.Map<Restaurant>(dto);
+        restaurant.CreatedById = userId;
         _dbContext.Restaurants.Add(restaurant);
         _dbContext.SaveChanges();
         return restaurant.Id;
